efs_shell: Allow ls when stat data is empty
When we have list permission for a directory but not read permission for
a file within that directory, all metadata for that file except its name
is zeroed out. Currently, this causes a crash looking up the "File type"
string. Add a new entry to gracefully handle this case. Unknown files
now look like this:

  +---------------------------------------------------------------------------------------------------------------------------+
  + File type | Special flags | File rights | File name               | File size | Modification        | Creation            +
  +---------------------------------------------------------------------------------------------------------------------------+
  + Unknown   |               | ---------   | 'mcfg'                  | 0         | 1969-12-31 19:00:00 | 1969-12-31 19:00:00 +
  +---------------------------------------------------------------------------------------------------------------------------+